chore: fix folder locations
