+++ packages/dashboard/src/components/design-system/Avatar.tsx
