+++ packages/dashboard/src/components/design-system/index.ts
@@ -1,0 +1,47 @@
+// Button
+export {
+  Button,
+  type ButtonProps,
+  type ButtonVariant,
+  type ButtonSize,
+} from "./Button";
+
+// Card
+export {
+  Card,
+  CardHeader,
+  CardBody,
+  CardFooter,
+  type CardProps,
+  type CardVariant,
+} from "./Card";
+
+// Breadcrumb
+export {
+  Breadcrumb,
+  type BreadcrumbProps,
+  type BreadcrumbItem,
+} from "./Breadcrumb";
+
+// Badge
+export {
+  Badge,
+  type BadgeProps,
+  type BadgeVariant,
+  type BadgeSize,
+} from "./Badge";
+
+// Avatar
+export { Avatar, type AvatarProps, type AvatarSize } from "./Avatar";
+
+// Skeleton
+export { Skeleton, type SkeletonProps } from "./Skeleton";
+
+// Tag
+export { Tag, type TagProps } from "./Tag";
+
+// Collapsible
+export { Collapsible, type CollapsibleProps } from "./Collapsible";
+
+// Spinner
+export { Spinner, type SpinnerProps } from "./Spinner";
