--- packages/dashboard/src/design-system/index.ts
@@ -1,43 +1,0 @@
-// Button
-export {
-  Button,
-  type ButtonProps,
-  type ButtonVariant,
-  type ButtonSize,
-} from "./components/Button";
-
-// Card
-export {
-  Card,
-  CardHeader,
-  CardBody,
-  CardFooter,
-  type CardProps,
-  type CardVariant,
-} from "./components/Card";
-
-// Breadcrumb
-export {
-  Breadcrumb,
-  type BreadcrumbProps,
-  type BreadcrumbItem,
-} from "./components/Breadcrumb";
-
-// Badge
-export {
-  Badge,
-  type BadgeProps,
-  type BadgeVariant,
-  type BadgeSize,
-} from "./components/Badge";
-
-// Avatar
-export { Avatar, type AvatarProps, type AvatarSize } from "./components/Avatar";
-
-// Skeleton
-export { Skeleton, type SkeletonProps } from "./components/Skeleton";
-
-// Tag
-export { Tag, type TagProps } from "./components/Tag";
-
-export * from "./components/Collapsible";
