+++ packages/dashboard/src/routes/layout.tsx
@@ -1,7 +1,7 @@
 import { component$, Slot, useStyles$ } from "@builder.io/qwik";
 import { routeLoader$ } from "@builder.io/qwik-city";
 import type { RequestHandler } from "@builder.io/qwik-city";
-import { Header } from "~/components/header/Header";
+import { Header } from "~/components/Header";
 
 // import styles from "./styles.css?inline";
